Fixing one set of review comments

+++ addons/controllers/antrea/antreaconfig_controller_test.go
@@ -4,29 +4,21 @@
 package controllers
 
 import (
-	"fmt"
 	"os"
 	"strings"
 	"time"
 
-	"github.com/vmware-tanzu/tanzu-framework/addons/pkg/util"
-
-	"github.com/vmware-tanzu/tanzu-framework/addons/pkg/constants"
-
-	"k8s.io/apimachinery/pkg/api/errors"
-
-	"sigs.k8s.io/cluster-api/util/secret"
-
-	cniv1alpha1 "github.com/vmware-tanzu/tanzu-framework/apis/cni/v1alpha1"
-
-	clusterapiv1beta1 "sigs.k8s.io/cluster-api/api/v1beta1"
-
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	v1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	clusterapiv1beta1 "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/vmware-tanzu/tanzu-framework/addons/pkg/constants"
+	"github.com/vmware-tanzu/tanzu-framework/addons/pkg/util"
 	"github.com/vmware-tanzu/tanzu-framework/addons/testutil"
+	cniv1alpha1 "github.com/vmware-tanzu/tanzu-framework/apis/cni/v1alpha1"
 )
 
 const (
@@ -36,7 +28,7 @@ const (
 
 var _ = Describe("AntreaConfig Reconciler", func() {
 	var (
-		clusterName             string
+		configCRName            string
 		clusterResourceFilePath string
 	)
 
@@ -48,9 +40,6 @@ var _ = Describe("AntreaConfig Reconciler", func() {
 		defer f.Close()
 		err = testutil.CreateResources(f, cfg, dynamicClient)
 		Expect(err).ToNot(HaveOccurred())
-
-		By("Creating kubeconfig for cluster")
-		Expect(testutil.CreateKubeconfigSecret(cfg, clusterName, "default", k8sClient)).To(Succeed())
 	})
 
 	AfterEach(func() {
@@ -59,21 +48,12 @@ var _ = Describe("AntreaConfig Reconciler", func() {
 		Expect(err).ToNot(HaveOccurred())
 		defer f.Close()
 		testutil.DeleteResources(f, cfg, dynamicClient, true)
-
-		By("Deleting kubeconfig for cluster")
-		key := client.ObjectKey{
-			Namespace: "default",
-			Name:      secret.Name(clusterName, secret.Kubeconfig),
-		}
-		s := &v1.Secret{}
-		Expect(k8sClient.Get(ctx, key, s)).To(Succeed())
-		Expect(k8sClient.Delete(ctx, s)).To(Succeed())
 	})
 
 	Context("Reconcile AntreaConfig for management cluster", func() {
 
 		BeforeEach(func() {
-			clusterName = "test-cluster-1"
+			configCRName = "test-cluster-1"
 			clusterResourceFilePath = "testcases/antrea-test-1.yaml"
 		})
 
@@ -84,8 +64,8 @@ var _ = Describe("AntreaConfig Reconciler", func() {
 				Name:      "test-cluster-1",
 			}
 
+			cluster := &clusterapiv1beta1.Cluster{}
 			Eventually(func() bool {
-				cluster := &clusterapiv1beta1.Cluster{}
 				err := k8sClient.Get(ctx, key, cluster)
 				if err != nil {
 					return false
@@ -101,7 +81,7 @@ var _ = Describe("AntreaConfig Reconciler", func() {
 				}
 
 				// Check owner reference
-				if len(config.Finalizers) == 0 || len(config.OwnerReferences) == 0 {
+				if len(config.OwnerReferences) == 0 {
 					return false
 				}
 				Expect(len(config.Finalizers)).Should(Equal(1))
@@ -146,7 +126,7 @@ var _ = Describe("AntreaConfig Reconciler", func() {
 			Eventually(func() bool {
 				secretKey := client.ObjectKey{
 					Namespace: "default",
-					Name:      fmt.Sprintf("%s-antrea-data-values", clusterName),
+					Name:      util.GenerateDataValueSecretNameFromAddonAndClusterNames(configCRName, constants.AntreaAddonName),
 				}
 				secret := &v1.Secret{}
 				err := k8sClient.Get(ctx, secretKey, secret)
@@ -160,9 +140,19 @@ var _ = Describe("AntreaConfig Reconciler", func() {
 				return true
 			}, waitTimeout, pollingInterval).Should(BeTrue())
 
+			Eventually(func() bool {
+				// Check status.secretRef after reconciliation
+				config := &cniv1alpha1.AntreaConfig{}
+				err := k8sClient.Get(ctx, key, config)
+				if err != nil {
+					return false
+				}
+				Expect(config.Status.SecretRef).Should(Equal(util.GenerateDataValueSecretNameFromAddonAndClusterNames(configCRName, constants.AntreaAddonName)))
+				return true
+			}, waitTimeout, pollingInterval).Should(BeTrue())
+
 		})
 
-		// TODO: Check status.secretRef after reconciliation
 		It("Should reconcile AntreaConfig deletion in management cluster", func() {
 
 			key := client.ObjectKey{
@@ -179,6 +169,13 @@ var _ = Describe("AntreaConfig Reconciler", func() {
 					}
 					return false
 				}
+
+				// Wait until the finalizer is added
+				if len(config.OwnerReferences) == 0 {
+					return false
+				}
+
+				// Delete antreaConfig
 				err = k8sClient.Delete(ctx, config)
 				if err != nil {
 					return false
@@ -189,7 +186,7 @@ var _ = Describe("AntreaConfig Reconciler", func() {
 			Eventually(func() bool {
 				secretKey := client.ObjectKey{
 					Namespace: "default",
-					Name:      fmt.Sprintf("%s-%s-data-values", clusterName, constants.AntreaAddonName),
+					Name:      util.GenerateDataValueSecretNameFromAddonAndClusterNames(configCRName, constants.AntreaAddonName),
 				}
 				secret := &v1.Secret{}
 				err := k8sClient.Get(ctx, secretKey, secret)
